+++ dbotconf/internal/generate_test.go
@@ -16,8 +16,12 @@ package internal
 
 import (
 	"bytes"
+	"fmt"
 	"io"
+	"path/filepath"
+	"runtime"
 	"strings"
+	"syscall"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -59,18 +63,25 @@ func newUpdate(pkgEco, dir string, labels []string) update {
 
 func TestBuildConfig(t *testing.T) {
 	root := "/home/user/repo"
+	if runtime.GOOS == "windows" {
+		// Ensure forward slashes on Windows and add drive letter.
+		// e.g. c:/home/user/repo
+		drive, err := syscall.FullPath("/")
+		require.NoError(t, err)
+		root = filepath.ToSlash(filepath.Join(drive, root))
+	}
 	mods := []*modfile.File{
-		{Syntax: &modfile.FileSyntax{Name: "/home/user/repo/go.mod"}},
-		{Syntax: &modfile.FileSyntax{Name: "/home/user/repo/a/go.mod"}},
-		{Syntax: &modfile.FileSyntax{Name: "/home/user/repo/b/go.mod"}},
+		{Syntax: &modfile.FileSyntax{Name: fmt.Sprintf("%s/go.mod", root)}},
+		{Syntax: &modfile.FileSyntax{Name: fmt.Sprintf("%s/a/go.mod", root)}},
+		{Syntax: &modfile.FileSyntax{Name: fmt.Sprintf("%s/b/go.mod", root)}},
 	}
 	dockerFiles := []string{
-		"/home/user/repo/",
-		"/home/user/repo/a/",
-		"/home/user/repo/b/",
+		fmt.Sprintf("%s/", root),
+		fmt.Sprintf("%s/a/", root),
+		fmt.Sprintf("%s/b/", root),
 	}
 	pipFiles := []string{
-		"/home/user/repo/requirements.txt",
+		fmt.Sprintf("%s/requirements.txt", root),
 	}
 
 	got, err := buildConfig(root, mods, dockerFiles, pipFiles)